Whole android example with gradle file
