+++ xapian-bindings/android/example/app/.gitignore
@@ -1,0 +1,2 @@
+/build
+*.aar
