+++ xapian-bindings/android/example/app/src/main/java/org/xapian/example/android/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.tj.xapiansample;
+package org.xapian.example.android;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
