+++ xapian-bindings/android/example/app/src/main/res/layout/activity_main.xml
