+++ xapian-bindings/android/example/app/src/main/res/values/strings.xml
