+++ xapian-bindings/android/example/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Sun Apr 03 23:09:13 IST 2016
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.10-all.zip
